Dynamically allocate WaveTables
After fixing #872, some users rightly complained about the expansion
in memory. This changes the Surge protocol to dyanmically allocate and
grow sample storage for wavetables, starting at a lower point than it did
before. Since this is a sensitive change I tested it extensively.

1. headless mac and linux plays every patch without a crash
2. valgrind linux shows no memory leaks
3. A wide variety of windows hosts both 32 and 64 bit tested

I'm going to ask for tests for this on various forums and consider
this plus the last correction as important ones for a candidate 1.6.1
early next week.

Closes #900