Surge UserInteractions stubbed with Zenity
Surge UserInteractions on Linux should really use
the vst event loop and vstgui and be hand written because
otherwise you have to popen zenity which can cause audio
pauses and pops.

But audio pauses and pops on anciliary actions are better than
nothing, so stick in a crudy zenity implementation which is
a start, and someone inspired to remove it at least has a spec
in the future.

Addresses #562
Closes #637 and #917