Refcount assert limit
One user, one time, when making a particularly wavetable heavy
set of patches, ran into the refcount limit on the critical
section. We aren't sure how and can't reproduce it. But if
refcounts are going to run away, they will run away to 10
just as easily as they run away to 5, so up the assert limit.

Closes #908