Musical TempoSync Values
temposync values used to have odd values like "2.667 1/4th notes"
and so on. We all kinda decided it was better to have musical terms
like "1/8 dotted" rather than "1.500 1/8" (or even better 3.000 1/16).
This change does that. Theres a small python script I used to figure
it all out, then the implementation as a member function in Parameter.

Closes #959