Another try to bring to front on Mac
Bring to front of the window for selecting directories worked
on some hosts but not in logic and live and so on. So clobber the
window order a little bit harder.

With this I now get the chooser window on top of surge in Logic 10.4.whatever
and bitwig 2.4.3 with the vst3

Closes #1001. Again. I hope!