Tuning Patches, Status Panel, and Patch Versions
This commit sets up a display of statuses in the synth so we
can bring to the front thigns like MPE and Tuning and bind the
menus more comprehensively. It also brings to bear the ability
(optionally) to store a tuning in a patch and the correct state
machine for what to do when loading a patch with tuning with
tuning active.

It almost completely addresses the remainder of #828
Addresses the streaming incompatability in #1035

Still outstanding is optionally a "lock" and to apply
the streaming version option to FX (which is #1037)

Handle streaming revision parameter set changes in OSCes. Deals with