Effects get a chance to react to streaming versions
With streaming version 11 (version 1.6.2) when I added parameters
to Vocoder and Sine we realized old patches had streamed nonsense
values for those new parameters. So add an opportunity along with
streaming version to reset my values to defaults.

To test this, load any of the vocoder patches in the system and see
that bands == 20 and so on.

Closes #1037