Add HiRes versions of some wavetables
Low sample count wavetables create high frequency noise in surge,
on purpose, as described in #830.

This commit, addressing part of #877, takes some of the more
super-low-sample-count wavetables, backs out the formulae used
to generate them, and creates a python script which lets us generate
them again with numpy.

Addresses but does not close #877 and #830