Add the LV2 plugin format (#1093)
Add the LV2 plugin format. 

This commit implements LV2 format on linux. There are some known issues still, including zoomability, but the plugin loads and runs in LV2 hosts.

The diff adds lv2 to build-linux just like other flavors; uses a new GitHub submodule to get the lv2 headers; ships inside the resulting deb file, and works in the CI pipeline.

To support this, we had to make some minor modifications to premake for mac, which was over-including software, and add an optional user data argument to the editor class. But besides that, almost all changes are isolated to new LV2-only code.