Surge++ First Commit
This commit is a massive rebase of the baconpaul/experimental branch
to start the surge++-master branch from the point where his 2 months
of slop work became something which other people could start co-developing.
Do this by squashing the hundreds of checkpoint commits and generating a
starting point for the surge++-master branch which we push to upstream.