OSC and LFO redraw on midi-cc slider move
If a slider maps to an LFO or OSC which is on display and
is midi cc learned, redraw the osc or lfo display just like
we do with param changes. Refactor the OSC check into a function
like I did for LFO (and should have done for OSC in the first place).

Closes #1232