Pixel-sensitive implied zoom in VST3
We want to make sure in onSize that we only reset our overall
zoom if the pixel size of the UI would have changed. We
were using a zoom factor change of 1% as a threshold before
but now we are using a zoom pixel change of > 1 pixel to
trigger resize on drag.

Closes #1219
Closes #1212