Support Release Velocity in the 16 vintage codebase
OK so I got a linnstrument, alright? Release velocity is super useful
and I should have done this forever ago. Basically plumb it through as
a modulation source in the DSP engine; hook it up properly in the VST3;
and modify the GUI so the Velocity modulator is an either-or modulator.
This results in the ability to modulate with release velocity just
as you can with anything else, albeit in a bit of a clumsy fashion.

Closes #811