Give clean message on invalid XML
Invalid XML used to cause a segfault. Now it causes a black 100x100 GUI
along with a clean Surge::UserInteractions::promptError. This change also
contains an incomplete slider parameterization, since that's what I was
working on when I screwed up the XML, but I want to get this into other devs
hands.