Temporary workarond for n_osc WT streaming (#1339)
Linked to #1337, the patch format assumes 3 oscillators per stream
and has binary offsets. With 6 oscillators per stream (that is using
n_osc rather than 3) things go wrong at unstream of 1.6.4 patches.
We need to fix this, but for now I made a couple of spots just have
'3' so we don't get segfaults.