Filter and FM Config; Readonly Nodes
The Filter and FM Config move over but in doing so we added a cool
new feature - a readonly node which can pair with a readwrite node
for alternate displays. In this case we just have an idealized SVG
display that doesn't paint yet, but will soon with images showing the
routings.