Implement Patch Wavetable Export
Given a Patch, it was impossible to save the wavetable as a
.wav which you could use elsewhere. You could copy and paste but
you couldn't extract. SO this adds code which exports the wav
into your library for you and lets you know, including doing smart
things with srge tags and the like so surge can use it in other contexts.

Closes #1341