Delay unstreams TempoSync correctly
In commit a1f187a trying to clean up uninitialized memory I left
in a specious override in the Delay only of temposync. That's
been there since April 2019 but meant that Delay temposync didn't
unstream! So remove that, since the temposync is initlaized in
Parameter::assign properly now, and now it unstreams.

Closes #1401