FrameClose on editor::close (#1430)
As noted in #1406 we don't handle the frame close cycle properly
which most DAWS compensate for, improperly probably, but which
FreeStyle rightly didn't, leading to a crash.

For now, before the holidays, put in this fix as VST2 only so we
can get some testing. But keep the issue open to do a careful review
of the close lifecycle in the editors in VST3 AU and maybe LV2 also.