Fix an Init Delay Timing Condition
Init on Delay with temposync could happen before the global
tempo was set when you unstreamed in a DAW. As a result the
lag for the time would catch up on the first and only first
play leading to glitches. To fix it, detect if setvars(true)
(the init condition) is called when temposync hasn't been set
yet and defer the initialization if so.

The frequency shifter would have the same bug so fix that also.

Closes #1444.