Support KBM Files (#1462)
Addresses #1041

This larger diff supports KBM files inside the surge engine
but doesn't include the workflow to load them in the UI or display
them at tuning time yet. Rather it sets up the ability to parse
KBM files and push them onto the storage and note architecture
and uses the Unit Test frameowrk to show that the synth is properly
retuned and remapped.

- Unit tests confirm we parse valid KBM files properly
- Add SurgeStorage hooks for a current mapping
- Confirm tuning and mapping for full-keyboard mappings works
- Allow users to pick a KBM file

Still work to do on persisting and showing.