Octave Shifts with Tunings
Should an octave be 12 notes? Or should an octave be the size
of your current scale? Surge had hardcoded 12 notes in a few
places because that was the size of the scale, but that's
counterintuitive with larger or smaller tunings (and also makes
layered sounds break). So go from 12 -> count. Add unit tests also.

Addresses #1473