VST2 Channel Split Selection Corrected (#1486)
The VST2 when setting a parameter sends automation which loops back
into the ::idle once more than the VST3 and AU do (and I think
improperly) but this meant that the scenemode override for mapping
hack was not applied properly. Fix that with further hacks. But they
are hacks you know.

Closes #1485