Fix MPE Pitch Bend with non-MPE devices
MPE pitch bend with non-MPE devices would double bend since the
(really unused) MPE global pitch bend got double counted. So make
it just be single counted, and put the fix in place with a regtest.
While at it, add a regtest for non-monotonic tunings and tuning in
the presence of pitch bend.

Closes #1489
Closes #1487