Add a "ParameterAlt" display for tooltip; use for Freq (#1515)
The addition of a note to the frequency display meant the percentage bounced
around like crazy. So add the capability to ahve an alternate parameter display
which is only used on the tooltip (not sent to the DAW/Automation) and left
aligns. Use that for the note approximation of frequency and for the position
in a wavetable.

Closes #447