BiQuad frequency should not be tuning aware
The BiQuad frequency calculation used the pitch table, but
not in the intent of the pitch of a note, rather as a computational
shortcut for 440 * 2^x/12. So when we retuned things like FX
cutoffs and the global HPF filter had their cutoff changed.

Fix this by adding a note_table_ignoring_tuning which is
always 440 * 2^x/12, using it in the biquad, and writing an
assertive unit test to amke sure the HPF is tuning invarient.

The note_to_pitch was used as a frequency conversion similarly
in the filter setup sections; so we need to adjust those similyarly
by adding an untuned omega.

Finally the FX with time used note-to-pitch to get a time lag
and aain this should be tuning independent

Closes #1545