Add Extended mode to Unison Range (#1562)
Unison was +/- 100cents or +/- 16hz. Add an extended range
so it is now +/- 1200 cents or 192 hz. With the 1200 cents and
uni count 3 you can get octaving really nicely using unison.

Disable the S*H Sample Random Test while in there too; it was
a bit flaky in the frequency detector

Closes #1552