LV2 State Adjustments
The LV2 takes the ports, not the synth, as authoritative
so some UI actions which change multiple values - notably
changing an FX or Oscillator or LFO Type - need to send
bulk parameter automation messages to make the LV2 consistent.

This may be useful/nice-to-have in the VST2/3/AU also but since
they capture values from the synth at all times, it won't make
an inconsistency. After the 1.6.6 release, go back to this diff
and consider activating it for the other hosts also, just
for consistency.

Addresses #1577