Towards a Surge SKIN Engine; Step 1 of many (#1620)
This is the first step towards a Surge skin engine. It contains
a variety of fundamental architectural changes to Surge UI layer and
additionally gives us the ability to do skins with different colors
and graphics as long as we are pixel perfect aligned still (moving
components comes soon).

The diff includes

- Giving proper names to surge parameters
- Generating and reading skin XML files from shared and user areas
- Switchable assets; functionally parameterized colors
- A dark-mode skin provided by @itsmedavep

This addresses #1597. I will use that as a master issue and break
out my forward todo list into more issues once this is merged.