Add a Skin-optional button glyph for status panel
The status panel button is drawn in code. Make it so that
you can also draw the button with a 2 state button. The button
is 31x11 so the sprite glyph is 31x22 with the off state first.
It is specified in a skin with <image id="STATUS_BUTTON" resource=".."/>

Addresses #1634