Add UI Instrumentation (#1703)
Add UI Instrumentation which is on in this version, and instrument
a few points looking for Vincey's slowdown/leak. While in there, form
the theory that windows RC loading was leaking in skin mode, so add
a (small memory static leaky) cache which has a leak of fixed size
once as opposed to resource size regularly.