Experiemnt replacing fmod
Phase limiting on super high FM and rate with the sin osc is still
a problem with our fastsin approx, running CPU high. An experiment shows
that fmod - especially in large ranges (mac) or in general (windows)
is just slow, so replace it with a more naive divide/mod which will
work for our expected range more quickly.

Addresses #2040