Macro Automation in Render Mode (#2134)
There's a few APIs to set a param in a VST3. setParamNormalized,
setParamAutomated, and processParameterChange(list). Funnily
Reaper calls setParamNormailzed in realtime playback mode and
processParameterChange(list) in render mode. But I had forgotten
to put in the special case for macros in processParameterChange
so they got dropped in reaper render. Fixed.

Closes #1999