Fix a Reverb2 Pre-Delay out of bounds
Reverb2 if you modulated pre-delay out of bounds very very far could
explode. Fix it in two ways. (1) if->while in wrap code in the delay
in case I miss a spot; (2) clamp the predelay to far less than the delay
line (to allow the diffusion to spread across the delay)

Closes #2208