Rework the Param -> String architecture
The Param -> String architecture was becoming unwieldy as we
implemented display modes and typeins for modulations and different
types with different contents. So instead of having lots of sprintfs
which differed only by a constant or a unit, add a DisplayInfo structure
to a param which tells it how to participate in stringification.
There's 4 types, Custom (which we don't really use for floats),
Linear, a 2^bx and decibel.

And do this with having memcpy only members in Parameter

This lets us move tons of stuff from code to config and generally clean things up.

A big change, but one which adds uniformity to the codebase so should
reduce bugs.