Found and fiixed more skin issues
Main GUI had a VRule that was wrong on the left side of filter configuration
Approached drawing the hover states differently. This was after Paul had reports of errors drawing on Linnux.
UUpdate for the R & L buttons buy cutoff and resonance sliders biuttons
Fixed ruler tiicks on LFO when temposynced.