PatchBrowser Infrastrcuture for SurgeXT; Disabled at Runtime
This commit starts us down the path of generating a patch browser.
It binds Surge to sqlite3, has a feature extractor -> SQL loader,
has a rudimentary juce table which displays the database, and
can in fact work. but right now it is off by default.

If you want to turn it on, the patchDb std::make_unique in SurgeStorage.cpp
and the menu addition in CPatchBrowser.cpp which are commented out, comment in,
restart surge, and voila. It will (every time) compeltely recreate your database
and let you browse it.

Infinitely more work to do here f course, but wanted to get this into the
mainline of xt-alpha so I don't have to keep rebasing this branch.

Addresses #2359