Resolve clicking bug with phaser, convert stage frequencies to hz
This should resolve a bug with a click sound at the onset of notes following a period of silence (caused by a difference in the way process_only_control(..) sets up the phaser stages compared to process(..)). It also converts the base cutoff frequencies and modulation spans of the allpass frequencies to hz for easier tweaking.