Alpha-level prototype of minimum viable MSEG editor (#2473)
An MSEG LFO Source is a lovely thing to imagine in Surge,
and luckily, now we have one. Or at least the start of one.
This commit is the minium viable product we could call an
alpha of the mseg. You can save msegs, add and remove points,
and so on. It has a warning that it is alpha. But enough
to merge into the nightly.

Addresses #982