Introduce in-source JUCE plugins
We've known that JUCE plugins which contain surge are possible
(see surge-fx; and sort of vcvrack) but they didn't build in-tree.
If we are going to consider moving Surge to JUCE fully we would
want in-tree JUCE working.

Luckily Juce6 is cmake and surge is cmake, so you can make this all
work with some cmake shenanigans, which this pull request provides,
along with a start of a SurgePatchPlayer plugin and standalone.