Integrate Airwindows Open Source plugins
The Airwindows open source plugins (https://github.com/airwindows/airwindows)
provide a variety of useful FX. This commit takes a collection of them and
provides the infrastructure so that they can be compiled into and loaded
into a Surge FX bank.

Airwindows is released under the MIT license and is used here after discussion
with the author https://github.com/airwindows/airwindows/issues/17

There's still some stuff to do here - notably #2509 will be required to
finish the UI in one spot - but this is a fully functioning effect.

Addresses #2507