Menu Controls for Sliders and LFO Type
This commit introduces a menu type control to replace a slider
which is ct_ controlled in SurgeGUIEditor and which uses the
underlying discrete slider mechanism. Applied to Airwindows,
Flanger, and a couple of other spots. Addresses #2509 but
still work to do.

While in there, I realized I could quickly make the LFO
menu work also, so did.

Closes #2233