Expand WT Format
The WT Format with integers (flag & 4 == 1) assumes int16
has range 0-2^15. This requires all sorts of normalization
and stuff. That has to stay, but if we add another flag
(flag & 8 == 1) we can then do the 2^16 -> 2^15 in code rather
than at WT time (and in the future, perhaps even keep that extra
bit of precision).