Move MiniEdit to VSTGUI loop
spawn_miniedit was an OS-implemented version of a simple popup
for a value which was blocking in code. Replace it with an async
callback based version which renders in VSTGUI consistently
across platforms.

Allows us to remove our libs/wtl dependency; and also remove
a large swath of windows specific and mac specific code.