Implement MIDI smoothing options
Midi smoothing was a never-finishing exponential creep
towards the target. This had a variety of undesirable
effects, including making controllers seem not very
snappy and making DAWs show creeping values.

Replace that with a set of options to have different
MIDI input smoothing curves and speeds, which are
a synth-wide default.

Closes #1412