Windows: Switch to Unicode
This makes Windows builds Unicode-aware, most notably unlocking the full
range of NTFS file system paths.

To avoid #ifdef, paths in std::strings are now UTF-8, like on other
platforms. They are converted to and from the native UTF-16 as needed.
This is mostly transparent, except when converting between fs::path and
std::string. There, the following should be observed:

- Use path_to_string() instead of .generic_path()
- Don't pass strings into fs::path API, convert them to paths first via
  string_to_path()

Otherwise, on Windows and if the path contains chars other than 7-bit
ASCII, either a system_error will be thrown, or (less likely) the path
could be mistranscoded.

This should become less and less of an issue as fs::path <-> std::string
conversions are refactored away.

Fixes #2579