Move to a Data Driven UI Layout
This substantial commit, in service of #2763 and the other
skin engine issues, moves the specification and layout
for the surge UI entirely to a set of data objects,
and moves the skin engine to provide overrides for those
objects. This materially completes many features of the
skin engine as far as layout, motion, and computing position.

It is documented more extensively in src/common/SkinModel.h

Among changes which happened here were:

1. Remove all positoin information from parameters at construct time
2. Add a set of (incomplete) tutorial skins
3. Add a full data description of the UI
4. Remove the CStatusPanel and use individual buttons
5. Re-work drag-n-drop to allow drop-anywhere and consolidate throuhg SGE

and more

Addresses #2763