An Interface for ID Management
For this commit, please see the comment in the new
SurgeSynthesizerIDManagement.cpp. This implements
a type system for IDs so we can cleanly distinguish
between synth IDs, DAW Ids, and have code which maps
between them cleanly, with well named methods, in one place.

Intent is to have this code be live in 1.8 then use it to
expand params in 1.9.

Closes #2493