AirWindows FX Lifecycle Changes
These changes make sure that the FX lifecycle allows
Airwindows to appropriately initialie itself in various
load, copy, drag, etc... paths. Basically it gives us a
post-update-via-load-event hook which resets us appropriatley
and assumes we are disabled, which will immediately be
changed if we are not. Avoids the various uninitialized
states which occured due to mis-interaction with streaming.

Closes #2661
Closes #2620