Patch an automation first-time problem in Live
First time you adjusted a switch in Live it would use
the old value, force that on, and flicker. I think this
is because of a surge bug I can't find, but for now, and
only in live, in controlBeginEdit make sure the value
is set before you release the DAW to ask it of you
(which again it only does in Live first time round, I think
when it is putting the param in the little xy box thing).

Closes #3283 but not very satisfactorily