Move Beginedit to point where value consistent in switch (#3477)
Live has a problem which we haven't diagnosed over in #3283 where
our beginEdit needs to set value from control less we get a peculiar
setValue from VST3. This is the beginEdit / VST3 / Live. To do that
the control needs the right value before beginEdit is called which
all our controls do excelyt a binary switch which basically made
solo do the wrong thing in live. This fix to move beginEdit to after
the value change accomplishes that.

Closes #3283 again