Correct parameter_normalized for bools
parameter_normalized for bools ignored the input value.
This meant Ardour automation didn't work, but was also
just plain wrong.

Along the way we thought it was the stepCount so I added
SS functions to allow us to do a public queriy of boolean
but didn't use them in 18. See #3577

Closes #3576