Rework the midi key modulators
The midi key modulators interacting with voice were a source of
constaint pain since the voice modes are optimizing for voice
management not midi state management. As a resul thtings like
"highest" on mono mode with priority mode at lowest didn't work.

So maintain an array of which key is activating which scene
(after all the channel resolution) with a count of midi number
and use that for highest.

Also add extensive regtests

Closes #3597