Invalidate OffScreen caches first idle after zoom (#3614)
In some DAWs you have a resize at open event not order quite properly
with the offscreen cache setup which meant, at 100% in live with the
royal skin on windows, you got a black bar on the background outside
where the classic skin was. There's probably all sorts of clever fixes
involving re-ordering the draws and stuff, but I took a more basic
approach of after any reloadSkin event, blow away my offscreen caches
and repaint on the first subsequent idle. This is just a teensy
CPU impact on most daws, but on live fixes the paint too.