Retain assigned modulation when moving FX to other slots (#3695)
FX Mod when you re-order FX was not carried. Turns out this
is a wee bit tricky because we need to queue it up but we have
the structures to do so. So when you swap, create a queue of which
params are modulated onto, and then unpack it after clearing in
the recreate.

Closes #3694