Add skin connector for positioning the Store Patch dialog (#3824)
* Add skin connector for positioning the Store Patch dialog
Prior to this PR, the Store Patch dialog had a fixed position and size. This PR introduces a new skin connector which will allow repositioning that dialog.
No size adjustment yet (but the code to do so is commented out and ready for introducing it later on if need be).