An implementation of mod smapping for A2^bx (#4058)
ModSnapping for a 2^bx is ambiguous. Do you want to snap so
yoru distance is integral steps? Do you want to snap evenly in
exponentiated or log space? All these things conflict. This
diff makes a choice: Snap evenly in units of the exponentiated
space so if you are modulating 524hz you modulate in steps of 10
or .25 seconds in steps of .01 or so.

While in there, add a not-unreasonable default for custom.

Addresses #1024.