Move MIDI Control Write from config.xml
Wierdly, the system would write MIDI config to configuration.xml
which mostly failed but sometimes worked. Now we read it if there's
no user config, but can write to the userdatapath in SurgeMIDIDefaults.xml
if auser chooses to save a new default with a menu action (as well
as existant DAW state save and stuff) and read that file preferentially
if it exists.