Move src/common/gui to src/gui (#4428)
As part of the XT refactoring, I moved the src/common/gui directory
out of common. This leaves common as the engine, basically, for
running the synth and allows us to start refactoring the synth/gui
interaction some.

If you are looking at git log in src/gui and thinking 'where did
that history go', you want to run with `git log --follow` and it
will go back to the history before XT also.