Modify the way we support AVX (#4480)
Blanket -mavx means that static initializers add avx intrinsics
and what not. Instead we want to use a conditional approapch through
splitting up libraries and only compiling the necessary program
units with -mavx as described in #4479.

We haven't done that yet, but we will. So even though this
Closes #4479, there's still work to be done when we use
AVX for the first time.