Creation of RPM package builder.
Added a script (make_rpm.sh) to support the ability to build an RPM on rpmbuild enabled systems.
Added a custom build rule (surge-xt-rpm) to CMakeLists to invoke the make_rpm script.
Added a build step to make_deb.sh to create a tarball of the same folder hierarchy used to build the deb package.
Modified call to find shared libraries to use -exec instead of piping to xargs as xargs was barfing on spaces. Alternatively, we could >
Added some additional build instructions to Readme.