CHSwitch2 to Surge::Widgets::MultiSwitch
The inexorable push to rewrite the lot of it in JUCE takes a big
step forward as the widely used CHSwitch2 becomes a pure juce
widget.

As I went I contnue to add helpers and commonly used APIs I need.
In this case, a wheel accumulation helper and also an API
on skin to get the full hover bundle set.

Addresses #4587