CSurgeSlider -> Surge::GUI::Widgets::ModulatableSlider (#4617)
This is a big change, and it's a big change for two reasons.
First, I mean, I rewrote the slider class.
But also, I gave it a proper C++ interface which other modulatable
display items can adapt to. So we can write knbos ,And pads,
and really whatever else you want.