Restore patchid on initial load
The initial load would ignore the patchid-from-name
guess since we had already found Init Saw. Make it so
we beleive the name with various checks to make sure
that is reasonable. Closes #4625.

Also puts the inti patch name and template as a member of
storage and uses that consistently, and allows a user default
for the patch to use there.  Closes #4502.