Greatly shrink the Synthesizer::ID code
Synthesizer::ID was a bunch of plubming to allow indices,
ids, and param ids to all co-exist, but iwth JUCE we really
only need param ids.

I could have removed this class altogether but decided to keep
the indirection until a rainy day, so the ID is now a simple wrapper
on an int which is the param index.

Closes #4866