Select patch check mark more reliably
If you have an identical category and patch name betwene
the user and factory everything works except the check mark.
With this change it works even more reliably with only one
unfixable case still off (that is, if you save a session with
the factory patch selected on restore it will think you are on the
user patch with the checkmark, even though the patch is fine).

Supress the error message now thigns are better

Closes #4668