Add User/Factory flag to default patch
Add User/Factory flag to default patch. Unfortunately it is rather
hard to set this to Factory if you ahve a duped category and patch
name but in that case it basically picks user, for the same reason
as all the other same-cat-same-name ambiguities.

So not perfect but closes #5058 in the same way we closed the other
collision problem.