Less steps for vkb vel adjustment, setWantsKeyboardFocus to false (#5157)
This makes the virtual keyboard always receive keypresses (when we allow it of course - so when typeins are focused, VKB is not doing anything). Doing this makes using QWERTY for virtual keyboard atually sane, since you don't have to refocus VKB every time you tweak a slider with mouse etc.