A collection of Tuning & Formula UI Updates (#5273)
This commit adds a bunch of Tuning and Formula UI updates
and gets most of the way to skinning these things (but not
all the way there). Along the way

1. Added a self draw multiswitch
2. Update the tuning from tabs to an mseg control strip
3. Update the formula from tabs to an mseg control strip
4. Add Tuning SCL/KBM syntax highlighting
5. Add start of skin colors in tuning
6. Make the Formula debug UI look less mickey mouse
7. Fix two standalone regressions with Tempo Initial Color and
   title bar contents

Woof! That's a lot and I'm a bit fried on it so let me get this
into the nightly and take a break.