A major rework of the Modulation List screen (Incomplete) (#5316)
This commit is a major rework of the look, feel, and action
of the modulation list screen. It is incomplete in several ways
including features being missing and drawing and layout sitll
being a placeholder requiring design, but critically it uses
surge components and skins, uses the design language of the other
overlays, coordinates wiht the synth properly, adds filter, sort, and
other capabilities, and fixes innumerable little bugs which were in
the prior version.

Still to do includes some efficiencies, the add, skinning and designing
the row, and more.