Formula and MSEG setup in SurgePatch ctor
FOrmula and MSEG states weren't set in SurgePatch ctor.
This is fine in almost all cases *except* when running
without a data dir. In that case we don't find init saw so
dont' stream xml so don't do the setup in the load xml path.

Do it here just to be safe for that edge case