Initialize LFOs to poast-end-of-envelope
LFOModulationSource::assign for some wierd reason initialized
the envelope state to 'half way through'. This didn't matter if
you sent attacks, but for keytrigger SLFO it meant that on patch
load the first time the LFOs were ringing until you pressed a key
to run the instance through the envelope. So start at the end
and then it all works

Closes #5359