Don't launch a popup if a popup is popped up (#5398)
As described in #5395, 1.9 won't launch a popup with a click
if a popup is open and juce will. This change guards the popups
so this won't happen (namely if there is a meny to dismiss dont show
the new popup). This commit is isolated so if we dislike that
behavior we can revert it easily.

Closes #5395