A Dirty Flag, but not much done with it
This adds a dirty flag to the patch which lets you see when you
have modified it. We currently don't *do* antyhing with the flag
other than put an * by the patch name, but in the future we can do
quite a lot.

But since we just show it this only addresses part of #733