Correctly use aspect ratio constrainer (#6005)
This change correctly sets the aspect ratio constrainer for the
juce resize handler.

In standalone it perfectly preserves it
In bitwig mac vst3 it perfectly preserves it
In reaper mac the window is embedded and does not obey the content
aspect ratio
In logic there is no drag to resize

So seems it at least makes it 'better or same'. We should test this
in the daws showing problems in #5211 (which I dont have access to
right now) to see if it helps.